Touchscreen works but needs a better structure

+++ src/xpcc/driver/ui/display/graphic_display.cpp
@@ -104,7 +104,7 @@ xpcc::GraphicDisplay::drawLine(int16_t x1, int16_t y1, int16_t x2, int16_t y2)
 			yStep = -1;
 		}
 		
-		for (uint8_t x = x1; x < x2; ++x)
+		for (uint16_t x = x1; x < x2; ++x)
 		{
 			if (steep) {
 				(this->*draw)(y, x);
@@ -124,7 +124,7 @@ xpcc::GraphicDisplay::drawLine(int16_t x1, int16_t y1, int16_t x2, int16_t y2)
 void
 xpcc::GraphicDisplay::drawHorizontalLine(glcd::Point start, uint8_t length)
 {
-	for (uint8_t i = start.getX(); i < (start.getX() + length); ++i) {
+	for (uint16_t i = start.getX(); i < (start.getX() + length); ++i) {
 		(this->*draw)(i, start.getY());
 	}
 }
@@ -132,7 +132,7 @@ xpcc::GraphicDisplay::drawHorizontalLine(glcd::Point start, uint8_t length)
 void
 xpcc::GraphicDisplay::drawVerticalLine(glcd::Point start, uint8_t length)
 {
-	for (uint8_t i = start.getY(); i < (start.getY() + length); ++i) {
+	for (uint16_t i = start.getY(); i < (start.getY() + length); ++i) {
 		(this->*draw)(start.getX(), i);
 	}
 }
